CppCoreCheck - Arithmetic Rules.

+++ BlakesEngine/Core/beHash.cpp
@@ -115,7 +115,7 @@ void MurmurHash3_x86_32 ( const void * key, int len,
   //----------
   // body
 
-  const uint32_t * blocks = (const uint32_t *)(data + nblocks*4);
+  const uint32_t * blocks = (const uint32_t *)(data + (size_t)nblocks*4);
 
   for(int i = -nblocks; i; i++)
   {
@@ -133,7 +133,7 @@ void MurmurHash3_x86_32 ( const void * key, int len,
   //----------
   // tail
 
-  const uint8_t * tail = (const uint8_t*)(data + nblocks*4);
+  const uint8_t * tail = (const uint8_t*)(data + (size_t)nblocks*4);
 
   uint32_t k1 = 0;
 
@@ -176,7 +176,7 @@ void MurmurHash3_x86_128 ( const void * key, const int len,
   //----------
   // body
 
-  const uint32_t * blocks = (const uint32_t *)(data + nblocks*16);
+  const uint32_t * blocks = (const uint32_t *)(data + (size_t)nblocks*16);
 
   for(int i = -nblocks; i; i++)
   {
@@ -205,7 +205,7 @@ void MurmurHash3_x86_128 ( const void * key, const int len,
   //----------
   // tail
 
-  const uint8_t * tail = (const uint8_t*)(data + nblocks*16);
+  const uint8_t * tail = (const uint8_t*)(data + (size_t)nblocks*16);
 
   uint32_t k1 = 0;
   uint32_t k2 = 0;
@@ -296,7 +296,7 @@ void MurmurHash3_x64_128 ( const void * key, const int len,
   //----------
   // tail
 
-  const uint8_t * tail = (const uint8_t*)(data + nblocks*16);
+  const uint8_t * tail = (const uint8_t*)(data + (size_t)nblocks*16);
 
   uint64_t k1 = 0;
   uint64_t k2 = 0;
